Change license to LGPL
If the code was GPL, it wouldn't be able to be linked with the user's
application without making the whole application public under the GPL.
This would be largely incompatible with what our government/commerical
customers are trying to do.  Change this to the LGPL so that they only
have to re-contribute their internal modifications to the Waveform API
itself.