Implement pthread work queues
Use pthread work queues to establish queues in which to run callbacks.  This way we don't
block our main command and control thread for user operations.  Also, user operations have
less of a chance of blocking each other.  We also might be able to have blocking command
calls that wait for a response because we can afford for a task to be blocked because
the workqueue implementation should start more threads as the old ones get gummed up.