Flip every sample
We now potentially have the RAW mode which will have valid data in both the members of the
sample pair.  Byte swap both of them before handing them off.  If we end up sending the whole
packet to the user, we may need to just byte swap the whole packet for them to make it
easy.