Provide for VITA-49 packets without timestamps
The metering code in smoothlake expects packets that
do not have timestamps, otherwise it will try to interpret
the timestamp as meter samples.  Create code to try to
handle gracefully packets with and without timestamps.

This implementation isn't immediately ideal and probably
requires some more evolution and thinking.  But it does
have the advantage of working correctly and provides
and incremental change that can be built upon.