Handle data callbacks on our own work queue
libpthread_wq wasn't being very deterministic about
execution order of the callbacks.  This was causing problems
for the frequency hopping guys such that they were getting
out of order packets, which are problematic.  This commit
implements our own simple work queue so that we can control
this a bit more.  As long as libevent gives us events in
order (which it should), then we should be able to
dispatch the callbacks in that order.