DF-7747 Restructure Packet Transmission
There were a few potential bugs with packet transmission and
it made more sense just to rewrite.  We now use an improved
algorithm for sending.

The event is created in the event loop that fires on the socket
being ready to accept reads.  We keep the event disarmed by
default.  When we get a packet from the user, we first check to
see if the event is armed, and if so, we add the packet to the
end of a queue to be sent by the event loop.  If the event loop
is disarmed, we merely try to send the packet directly to the
socket.  If it succeeds, everything is copacetic and we return to
the user.  If it fails because the socket buffer is full
(EAGAIN or EWOULDBLOCK), we queue the packet and arm the event
so that the event loop will run the queue when the socket
becomes available.

When the event loop empties the queue, it disarms itself so that
direct sends continue unmolested.

This should allow us to queue packets when things get busy, but
still send packets direct when the kernel can keep up, which
should be the vast majority of the time.